test: 자신의 키워드 조회 api 테스트
