+++ src/test/java/in/koreatech/koin/acceptance/CommunityApiTest.java
@@ -10,11 +10,8 @@ import org.springframework.http.HttpStatus;
 
 import in.koreatech.koin.AcceptanceTest;
 import in.koreatech.koin.domain.community.articles.model.Article;
-import in.koreatech.koin.domain.community.keywords.model.ArticleKeywordUserMap;
 import in.koreatech.koin.domain.community.articles.model.Board;
 import in.koreatech.koin.domain.community.articles.model.Comment;
-import in.koreatech.koin.domain.community.keywords.repository.ArticleKeywordRepository;
-import in.koreatech.koin.domain.community.keywords.repository.ArticleKeywordUserMapRepository;
 import in.koreatech.koin.domain.community.articles.repository.ArticleRepository;
 import in.koreatech.koin.domain.community.articles.repository.CommentRepository;
 import in.koreatech.koin.domain.user.model.Student;
@@ -23,7 +20,6 @@ import in.koreatech.koin.fixture.BoardFixture;
 import in.koreatech.koin.fixture.UserFixture;
 import in.koreatech.koin.support.JsonAssertions;
 import io.restassured.RestAssured;
-import io.restassured.http.ContentType;
 
 @SuppressWarnings("NonAsciiCharacters")
 class CommunityApiTest extends AcceptanceTest {
@@ -33,12 +29,6 @@ class CommunityApiTest extends AcceptanceTest {
 
     @Autowired
     private CommentRepository commentRepository;
-
-    @Autowired
-    private ArticleKeywordRepository articleKeywordRepository;
-
-    @Autowired
-    private ArticleKeywordUserMapRepository articleKeywordUserMapRepository;
 
     @Autowired
     private UserFixture userFixture;
@@ -532,93 +522,5 @@ class CommunityApiTest extends AcceptanceTest {
                        "current_page": 1
                    }
                 """);
-    }
-
-    @Test
-    @DisplayName("알림 키워드를 추가한다.")
-    void 알림_키워드_추가() {
-        String token = userFixture.getToken(student.getUser());
-
-        var response = RestAssured
-            .given()
-            .header("Authorization", "Bearer " + token)
-            .contentType(ContentType.JSON)
-            .body("""
-                {
-                    "keyword": "장학금"
-                }
-                """)
-            .when()
-            .post("/articles/keyword")
-            .then()
-            .statusCode(HttpStatus.OK.value())
-            .extract();
-
-        JsonAssertions.assertThat(response.asPrettyString())
-            .isEqualTo("""
-                {
-                  "id": 1,
-                  "keyword": "장학금"
-                }
-                """);
-    }
-
-    @Test
-    @DisplayName("알림 키워드를 추가한다. - 10개 넘어가면 400에러 반환")
-    void 알림_키워드_10개_넘게_추가시_에러() {
-        String token = userFixture.getToken(student.getUser());
-
-        for (int i = 0; i < 10; i++) {
-            RestAssured
-                .given()
-                .header("Authorization", "Bearer " + token)
-                .contentType(ContentType.JSON)
-                .body(String.format("""
-                        {
-                        "keyword": "keyword%d"
-                        }
-                    """, i))
-
-                .when()
-                .post("/articles/keyword")
-                .then()
-                .statusCode(HttpStatus.OK.value());
-        }
-
-        RestAssured
-            .given()
-            .header("Authorization", "Bearer " + token)
-            .contentType(ContentType.JSON)
-            .body("""
-                {
-                    "keyword": "장학금"
-                }
-                """)
-            .when()
-            .post("/articles/keyword")
-            .then()
-            .statusCode(HttpStatus.BAD_REQUEST.value());
-    }
-
-    @Test
-    @DisplayName("알림 키워드를 삭제한다.")
-    void 알림_키워드_삭제() {
-        String token = userFixture.getToken(student.getUser());
-        ArticleKeywordUserMap articleKeywordUserMap = articleFixture.키워드1("수강 신청", student.getUser());
-
-        var response = RestAssured
-            .given()
-            .header("Authorization", "Bearer " + token)
-            .pathParam("id", articleKeywordUserMap.getId())
-            .contentType(ContentType.JSON)
-            .when()
-            .delete("/articles/keyword/{id}")
-            .then()
-            .statusCode(HttpStatus.NO_CONTENT.value())
-            .extract()
-            .asString();
-
-        assertThat(articleKeywordUserMapRepository.findById(articleKeywordUserMap.getId()).isEmpty());
-        assertThat(articleKeywordRepository.findById(articleKeywordUserMap.getArticleKeyword().getId()).isEmpty());
     }
 }
